arreglado un fallo que tenia al cambiar de pregunta secreta y ahora creo que si que he terminado con los usuarios

+++ usuarios/eliminar.php
@@ -10,16 +10,26 @@
         } 
     }
 	if((!isset($_SESSION['email'])) && (empty($_SESSION['email']))) {
-		header('Location: ../login.php');
+		header("Location: ../login.php");
 	}
     require '../funcion.php';
     $consulta = new Consultas;
     if(isset($_GET['email'])) {
-        $email = $consulta->escapar($_GET['email']);
-        $consulta->preparar('DELETE FROM users WHERE email = ?', $email, 's');
+        if ($_GET['email'] === $_SESSION['email']) {
+            $consulta->cerrar();
+            $_SESSION['error'] = "<div class=\"alert alert-danger alert-dismissible fade show\" role=\"alert\">  <button type=\"button\" class=\"close\" data-dismiss=\"alert\" aria-label=\"Close\"><span aria-hidden=\"true\">&times;</span></button><strong>Error </strong>No puedes borrarte a ti mismo</div>";        } else {
+            $email = $consulta->escapar($_GET['email']);
+            $consulta->preparar('DELETE FROM users WHERE email = ?', $email, 's');
+        }
+
     } elseif (isset($_POST['checkbox'])) {
          foreach ($_POST['checkbox'] as $email) {
-            $consulta->preparar('DELETE FROM users WHERE email = ?', $email, 's');  
+            if ($email === $_SESSION['email']) {
+                $consulta->cerrar();
+                $_SESSION['error'] = "<div class=\"alert alert-danger alert-dismissible fade show\" role=\"alert\">  <button type=\"button\" class=\"close\" data-dismiss=\"alert\" aria-label=\"Close\"><span aria-hidden=\"true\">&times;</span></button><strong>Error </strong>No puedes borrarte a ti mismo</div>";
+            } else {
+                $consulta->preparar('DELETE FROM users WHERE email = ?', $email, 's');  
+            }
         }
     }
     $consulta->cerrar();
